+++ src/entity/root-entity.hpp
