+++ src/world.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include "box2d-debug.hpp"
-#include "entity.hpp"
+#include "entity/entity.hpp"
+#include "entity/root-entity.hpp"
 #include "inputmanager.hpp"
-#include "root-entity.hpp"
 #include "time.hpp"
 
 #include <SFML/Graphics/RenderWindow.hpp>
